pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wikigen"
version = "0.3.0"
description = "Wiki's for nerds, by nerds"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    {name = "Mithun Ramesh"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup",
]
dependencies = ['pocketflow>=0.0.1', 'pyyaml>=6.0', 'requests>=2.28.0', 'gitpython>=3.1.0', 'google-genai>=1.9.0', 'pathspec>=0.11.0', 'keyring>=24.0.0', 'mcp>=1.19.0', 'faiss-cpu>=1.7.4', 'sentence-transformers>=2.2.0', 'numpy>=1.24.0']

[project.scripts]
wikigen = "wikigen.cli:main"

[project.urls]
Homepage = "https://github.com/usesalt/wikigen"
Repository = "https://github.com/usesalt/wikigen"
Issues = "https://github.com/usesalt/wikigen/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["wikigen*"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince212",
    "ignore::DeprecationWarning:google.genai",
]

